pyproject.toml
[tool.poetry]
name = "teampass"
version = "0.1.0"
description = ""
authors = ["ycalk <alosev752@gmail.com>"]
packages = [
    {include = "teampass/entrypoint",from = "bases"},
    {include = "teampass/migrator",from = "bases"},
    {include = "teampass/database",from = "components"},
    {include = "teampass/logging",from = "components"},
    {include = "teampass/user",from = "components"},
    {include = "teampass/team",from = "components"},
    {include = "teampass/domain_core",from = "components"},
    {include = "teampass/admin_panel",from = "bases"},
    {include = "teampass/admin",from = "components"},
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pydantic-settings = "^2.13.1"
fastapi = {extras = ["standard"], version = "^0.135.1"}
dishka = {extras = ["fastapi"], version = "^1.9.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.48"}
alembic = "^1.18.4"
asyncpg = "^0.31.0"
structlog = "^25.5.0"
opentelemetry-api = "^1.40.0"
opentelemetry-sdk = "^1.40.0"
opentelemetry-exporter-otlp = "^1.40.0"
opentelemetry-instrumentation-sqlalchemy = "^0.62b0"
opentelemetry-instrumentation-fastapi = "^0.62b0"
aio-pika = "^9.6.2"
email-validator = "^2.3.0"
argon2-cffi = "^25.1.0"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pyjwt = "^2.12.1"
sqladmin = {extras = ["full"], version = "^0.24.0"}
opentelemetry-instrumentation-starlette = "^0.62b0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
alembic = "teampass.migrator:run"
entrypoint = "teampass.entrypoint:run"
admin_panel = "teampass.admin_panel:run"